Testcafe: add "disposeWidget" helper method.
