+++ packages/devextreme/testing/testcafe/helpers/createWidget.ts
@@ -1,98 +1,47 @@
 import { ClientFunction } from 'testcafe';
-import type { Properties as DataGridProperties } from '../../../js/ui/data_grid';
-import type { Properties as FilterBuilderProperties } from '../../../js/ui/filter_builder';
+import type { WidgetName, WidgetOptions } from './widgetTypings';
 
-export type WidgetName =
-'dxAccordion'
-| 'dxAutocomplete'
-| 'dxGallery'
-| 'dxButtonGroup'
-| 'dxCalendar'
-| 'dxCalendarView'
-| 'dxCheckBox'
-| 'dxColorBox'
-| 'dxDropDownButton'
-| 'dxDraggable'
-| 'dxTabPanel'
-| 'dxForm'
-| 'dxFilterBuilder'
-| 'dxSelectBox'
-| 'dxScrollable'
-| 'dxScrollView'
-| 'dxMultiView'
-| 'dxPivotGrid'
-| 'dxPivotGridFieldChooser'
-| 'dxDataGrid'
-| 'dxTreeList'
-| 'dxPager'
-| 'dxRadioGroup'
-| 'dxScheduler'
-| 'dxTabs'
-| 'dxTagBox'
-| 'dxContextMenu'
-| 'dxDropDownMenu'
-| 'dxChart'
-| 'dxMenu'
-| 'dxPopup'
-| 'dxPopover'
-| 'dxSpeedDialAction'
-| 'dxSortable'
-| 'dxButton'
-| 'dxTextBox'
-| 'dxTextArea'
-| 'dxToolbar'
-| 'dxTreeView'
-| 'dxDateBox'
-| 'dxDateRangeBox'
-| 'dxLookup'
-| 'dxOverlay'
-| 'dxList'
-| 'dxHtmlEditor'
-| 'dxNumberBox'
-| 'dxValidator'
-| 'dxFileUploader'
-| 'dxSlider'
-| 'dxDropDownBox'
-| 'dxFileManager'
-| 'dxSwitch';
-
-interface WidgetOptions {
-  dxDataGrid: DataGridProperties;
-  dxFilterBuilder: FilterBuilderProperties;
-  // todo write other widgets
+export interface CreateWidgetOptions {
+  disableFxAnimation: boolean;
 }
 
-export default async function createWidget<TWidgetName extends WidgetName>(
-  componentName: TWidgetName,
-  componentOptions: TWidgetName extends keyof WidgetOptions
+const DEFAULT_SELECTOR = '#container';
+const DEFAULT_OPTIONS: CreateWidgetOptions = {
+  disableFxAnimation: true,
+};
+
+export const createWidget = async<TWidgetName extends WidgetName>(
+  widgetName: TWidgetName,
+  widgetOptions: TWidgetName extends keyof WidgetOptions
     ? WidgetOptions[TWidgetName] | (() => WidgetOptions[TWidgetName])
     : unknown,
-  selector = '#container',
-  options: {
-    disableFxAnimation: boolean;
-  } = {
-    disableFxAnimation: true,
+  selector = DEFAULT_SELECTOR,
+  { disableFxAnimation } = DEFAULT_OPTIONS,
+): Promise<void> => ClientFunction(
+  () => {
+    (window as any).DevExpress.fx.off = disableFxAnimation;
+    const options = typeof widgetOptions === 'function' ? widgetOptions() : widgetOptions;
+    (window as any).widget = ($(`${selector}`) as any)[widgetName](options)[widgetName]('instance');
   },
-): Promise<void> {
-  await ClientFunction(() => {
-    (window as any).DevExpress.fx.off = options.disableFxAnimation;
-  }, {
+  {
     dependencies: {
-      options,
+      widgetName,
+      widgetOptions,
+      selector,
+      disableFxAnimation,
     },
-  })();
+  },
+)();
 
-  await ClientFunction(
-    () => {
-      const widgetOptions = typeof componentOptions === 'function' ? componentOptions() : componentOptions;
-      (window as any).widget = ($(`${selector}`) as any)[componentName](widgetOptions)[componentName]('instance');
-    },
-    {
-      dependencies: {
-        componentName,
-        componentOptions,
-        selector,
-      },
-    },
-  )();
-}
+export const disposeWidget = async <TWidgetName extends WidgetName>(
+  widgetName: TWidgetName,
+  selector = DEFAULT_SELECTOR,
+): Promise<void> => ClientFunction(() => {
+  ($(`${selector}`) as any)[widgetName]('dispose');
+  (window as any).widget = undefined;
+}, {
+  dependencies: {
+    widgetName,
+    selector,
+  },
+})();
