+++ packages/devextreme/testing/testcafe/helpers/widgetTypings.ts
@@ -1,0 +1,62 @@
+import type { Properties as DataGridProperties } from '../../../js/ui/data_grid';
+import type { Properties as FilterBuilderProperties } from '../../../js/ui/filter_builder';
+
+export type WidgetName =
+  'dxAccordion'
+  | 'dxAutocomplete'
+  | 'dxGallery'
+  | 'dxButtonGroup'
+  | 'dxCalendar'
+  | 'dxCalendarView'
+  | 'dxCheckBox'
+  | 'dxColorBox'
+  | 'dxDropDownButton'
+  | 'dxDraggable'
+  | 'dxTabPanel'
+  | 'dxForm'
+  | 'dxFilterBuilder'
+  | 'dxSelectBox'
+  | 'dxScrollable'
+  | 'dxScrollView'
+  | 'dxMultiView'
+  | 'dxPivotGrid'
+  | 'dxPivotGridFieldChooser'
+  | 'dxDataGrid'
+  | 'dxTreeList'
+  | 'dxPager'
+  | 'dxRadioGroup'
+  | 'dxScheduler'
+  | 'dxTabs'
+  | 'dxTagBox'
+  | 'dxContextMenu'
+  | 'dxDropDownMenu'
+  | 'dxChart'
+  | 'dxMenu'
+  | 'dxPopup'
+  | 'dxPopover'
+  | 'dxSpeedDialAction'
+  | 'dxSortable'
+  | 'dxButton'
+  | 'dxTextBox'
+  | 'dxTextArea'
+  | 'dxToolbar'
+  | 'dxTreeView'
+  | 'dxDateBox'
+  | 'dxDateRangeBox'
+  | 'dxLookup'
+  | 'dxOverlay'
+  | 'dxList'
+  | 'dxHtmlEditor'
+  | 'dxNumberBox'
+  | 'dxValidator'
+  | 'dxFileUploader'
+  | 'dxSlider'
+  | 'dxDropDownBox'
+  | 'dxFileManager'
+  | 'dxSwitch';
+
+export interface WidgetOptions {
+  dxDataGrid: DataGridProperties;
+  dxFilterBuilder: FilterBuilderProperties;
+  // TODO: write other widgets
+}
